requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-extensions==2.1.6
django-filter==2.1.0
django-heroku==0.3.1
django-taggit==1.1.0
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
gunicorn==19.9.0
idna==2.8
Markdown==3.1
psycopg2==2.8
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
whitenoise==4.1.2
